requirements.txt
browser-use
python-dotenv
requests
playwright

